finished edit and delete features for the admin & their styling, reponsive
